FIPS: Allow krad UDP/TCP localhost connection
libkrad allows to establish connections only to UNIX socket in FIPS
mode, because MD5 digest is not considered safe enough to be used for
network communication. However, FreeRadius requires connection on TCP or
UDP ports.

This commit allows TCP or UDP connections in FIPS mode if destination is
localhost.

Resolves: rhbz#2082189
Last-updated: krb5-1.22.1
Forward-ported-by: Andreas Schneider <asn@redhat.com>